translation units, dyn-array function
