+++ strelyaev.roman/P2/matrix.h
@@ -3,5 +3,7 @@
 #include <istream>
 void sMatrix (std::istream & in, int * a, int e);
 void dMatrix (std::istream & in, int * a, int e);
+void matrixSubtraction(int * original, int * reverse, int n);
+void clockwise(int * a, int rows, int columns);
 #endif
